Feature Update
Updated README.
Added BraceletActivity.
All buttons should now work as intended.
Tasks can now be added to the list and shared between different
activities, i.e. tasks added to the app will appear in the time chunk
screen.
Added in menu buttons for "add" and "bracelet", not images are not
final.
A signed version of the app is now included. This should work on Android
2.2+ but has only been tested on Android 4.4.2